+++ rotkehlchen/globaldb/manual_price_oracles.py
@@ -93,7 +93,11 @@ class ManualCurrentOracle(CurrentPriceOracleInterface):
                 if current_to_asset == main_currency:
                     return current_price, True
 
-        current_to_asset_price, _, used_main_currency = Inquirer.find_price_and_oracle(
+        # we call _find_price to avoid catching the recursion error at `find_price_and_oracle`.
+        # ManualCurrentOracle does a special handling of RecursionError using
+        # `coming_from_latest_price` to detect recursions on the manual prices and break
+        # it to continue to the next oracle.
+        current_to_asset_price, _, used_main_currency = Inquirer._find_price(
             from_asset=current_to_asset,
             to_asset=to_asset,
             coming_from_latest_price=True,